requirements.txt
flask==2.3.3
werkzeug==2.3.7
mysql-connector-python==8.0.26
python-dotenv==0.19.0
